Revert "Only filter out subsettings with lower scores."
This reverts commit 8acb8d8cc2fa037c51cf3ca1365ba06db85d5b16.

Reason for revert: Tentative fix for CrOS bots

Original change's description:
> Only filter out subsettings with lower scores.
>
> The launcher settings search currently filters out all sub-settings or
> sub-sections which contain an ancestor in the results page. This is
> sometimes bad because the specific setting is actually the desired
> result (see the crbug for an example).
>
> This CL changes filtering so that only sub-settings with an equal or
> worse score than their ancestor get filtered out.
>
> Tested on device with some common queries to check that results are
> still good.
>
> Bug: 1267815
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3320655
> Commit-Queue: Rachel Wong <wrong@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#949385}

Bug: 1267815
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3325294
Auto-Submit: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#949801}

+++ chrome/browser/ui/app_list/search/os_settings_provider.h
@@ -42,7 +42,7 @@ class OsSettingsResult : public ChromeSearchResult {
  public:
   OsSettingsResult(Profile* profile,
                    const chromeos::settings::mojom::SearchResultPtr& result,
-                   double relevance_score,
+                   float relevance_score,
                    const gfx::ImageSkia& icon,
                    const std::u16string& query);
   ~OsSettingsResult() override;
